refactor: header의 버튼을 ui 컴포넌트로 분리
