+++ client/src/components/Avatar.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function Avatar({ user }) {
+  return <button>{user.id}님</button>;
+}
